+++ src/hooks/data/utils.ts
@@ -1,9 +1,7 @@
 import moment from "moment";
 import groupBy from "lodash/groupBy";
-import mapValues from "lodash/mapValues";
 
 import type { CaseData, DataDict } from "../../types";
-import type { ApexOptions } from "apexcharts";
 
 export const map = <T, K extends keyof T>(array: T[], key: K) =>
   array.map((element) => element[key]);
@@ -63,75 +61,4 @@ export const calcDataForUS = (caseDataRowsByState: CaseData[]) => {
   const dateRowsUS = Object.keys(dateDataDict);
 
   return { dateRowsUS, totalCasesRowsUS, newCasesRowsUS };
-};
-
-export const getNumberWithCommas = (num: number) => {
-  return num.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
-};
-
-const axisOptions: ApexOptions = {
-  xaxis: {
-    labels: {
-      rotate: -45,
-      rotateAlways: true,
-      formatter: function (value: string) {
-        return value.replace("2020-", "").replace("-", "/");
-      },
-    },
-  },
-};
-
-const chartOptions: ApexOptions = {
-  chart: {
-    id: "basic-bar",
-  },
-
-  dataLabels: {
-    formatter: getNumberWithCommas,
-    textAnchor: "end",
-    offsetY: 20,
-  },
-  plotOptions: {
-    bar: {
-      horizontal: false,
-      columnWidth: "80%",
-      dataLabels: {
-        position: "bottom",
-        maxItems: 100,
-        hideOverflowingLabels: true,
-        orientation: "vertical",
-      },
-      colors: {
-        ranges: [
-          {
-            from: 0,
-            to: Infinity,
-            color: "#D9534F",
-          },
-        ],
-      },
-    },
-  },
-};
-
-export const makeChartData = (
-  name: string = "US",
-  categories: string[],
-  data: number[]
-) => {
-  return {
-    options: {
-      ...chartOptions,
-      xaxis: {
-        ...axisOptions.xaxis,
-        categories,
-      },
-    },
-    series: [
-      {
-        name,
-        data,
-      },
-    ],
-  };
 };
